pylib/requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.6b0
boto3==1.17.105
botocore==1.20.105
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
more-itertools==8.8.0
moto==2.0.10
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
praw==7.3.0
prawcore==2.2.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.1
requests==2.25.1
requests-oauthlib==1.3.0
responses==0.13.3
rsa==4.7.2
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==2.0.1
xmltodict==0.12.0
